Module deathframe_core::systems::prelude

Structs

ConfineEntitiesSystem
This system confines all entities with Transform and Confined to the rect defined in Confined, taking Size into account.

EntityLoaderSystem
The EntityLoaderSystem handles the loading and unloading of entities. Entities with the Loader component load entities when they are in range with Loadable entities, and Loadable entities are unloaded when no Loader entities are in range.

FollowSystem

InputManagerSystem
Handles all the logic for InputManager.

ScaleSpritesSystem
This system gets all entities with Transform, Size, SpriteRender, and ScaleOnce, and scales their sprite to their entity's size once; after scaling, the ScaleOnce component is removed from the entity.